implement: Corsの設定wおクラスに分離する。
