+++ youkai-server/src/main/java/com/github/kazuhitom/youkai/server/CorsConfig.java
@@ -1,27 +1,39 @@
 package com.github.kazuhitom.youkai.server;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
 
-import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Stream;
 
-//@Configuration
+@Configuration
 public class CorsConfig {
     @Bean
     public CorsFilter corsFilter() {
         CorsConfiguration config = new CorsConfiguration();
-        config.setAllowedOrigins(Arrays.asList("http://localhost:8080"));
+        config.setAllowedOrigins(allowOriginsList());
         config.setAllowCredentials(true);
-        config.addAllowedMethod("*");
+        config.setAllowedMethods(List.of("GET", "POST"));
         config.addAllowedHeader("*");
         config.addExposedHeader("Set-Cookie");
 
-        UrlBasedCorsConfigurationSource configSource = new UrlBasedCorsConfigurationSource();
-        configSource.registerCorsConfiguration("/**", config);
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/api/**", config);
 
-        return new CorsFilter(configSource);
+        return new CorsFilter(source);
+    }
+
+    @Value("${management.endpoints.web.cors.allowed-origins}")
+    private String allowOrigins;
+
+    private List<String> allowOriginsList() {
+        return Stream.of(allowOrigins.split(","))
+                .map(String::trim)
+                .toList();
     }
 }
